17 april 2018

+++ src/main/java/com/bridgelab/todo/notes/dao/INotesDao.java
@@ -25,7 +25,7 @@ public interface INotesDao {
 
 
 	Notes getNoteById(long noteId);
-	 List<Notes> getAllNotesByUserId(int userId);
+	 List<Notes> getAllNotesByUserId(long userId);
